Navigation - create simple navigation  in VkNewsMainScreenNew.kt without navGraph and adding to backstack
