+++ app/src/main/java/com/example/composeproject/MainActivity.kt
@@ -6,9 +6,6 @@ import androidx.activity.compose.setContent
 import androidx.activity.viewModels
 import com.example.composeproject.ui.theme.ComposeProjectTheme
 import com.example.composeproject.ui.theme.MainScreenNew
-import com.example.composeproject.ui.theme.MainViewModel
-import com.example.composeproject.ui.theme.TestLazyColumn
-import com.example.composeproject.ui.theme.VkNewsMainScreenViewModel
 
 class MainActivity : ComponentActivity() {
