+++ app/src/main/java/com/example/composeproject/TestComposableFunctions/ScaffoldTest.kt
@@ -1,4 +1,4 @@
-package com.example.composeproject.ui.theme
+package com.example.composeproject.TestComposableFunctions
 
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.height
