+++ app/src/main/java/com/example/composeproject/TestComposableFunctions/VkNewsMainScreenViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.composeproject.ui.theme
+package com.example.composeproject.TestComposableFunctions
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
